requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.1.3
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
